Adding API and UI implementation
- I tried to go a more 'DIY' approach after what you asked me, but I just couldn't do it. I watched a lot of videos on attaching SQL, and a few on just connecting to external server stuff (and so forth), but it all looked far to time consuming. This is 20% of my grade, and all of it is write-up. I can just embellish this part of the write-up and make it look more impressive than it actually is. I have like 3 months left, and I'd much rather spend it perfecting the 80% instead of being stagnant on this project.

If I do the write-up for what I have, and get to a point in studying where i'm extremely confident, and we're convinced that doing this part from scratch would significantly boost my grade, which I doubt, then if we have time i'm happy to come back to revisit it. But I'm becoming incredibly sluggish when it comes to doing anything to do with this project now, which I know isn't the best attitude, but I just don't hold any enthusiasm for it. <- Although that shouldn't reflect on my attitude for the subject as a whole. For that i'm still pumped. Just this server/project part.